requirements.txt
pytest>=7.0.0
openai>=1.0.0
cryptography>=42.0.0
pymupdf>=1.24.0
pdfplumber>=0.11.0
python-dotenv>=1.0.1
pydantic>=2.0.0
rich>=13.0.0
python-docx>=1.1.2
sentence-transformers>=3.0.0
jinja2>=3.1.3

# Segurança e Autenticação Web (SEC-005, SEC-006)
cryptography>=42.0.0
PyJWT>=2.8.0
Flask-Limiter>=3.7.0

# LLM SDKs
openai>=1.14.0
Flask>=3.0.0
gunicorn>=23.0.0
